Fix path to .gitmessage file in commit.template config
The `commit.template` config value is interpreted as a path relative to
the repository's root directory. Before, this config value stored the
path `.git/.gitmessage` in all cases. This is the correct path to the
`.gitmessage` file for normal repositories. However, if the repository
is a submodule of another repository, the `.gitmessage` file will be
stored in the `modules` subdirectory of the parent repository's `.git`
directory.

The solution here is to correctly compute a path that, when resolved
from the proejct's root directory, will resolve to the `.gitmessage`
file in all cases.